Added card rendering
Added to filter.html - "rendered-cards" block to filters block, because according to the layout, they go in one block.
Added and change styles for the cards in card.css and filters.css.
Finally added a rendering  for the cards by queries.

In progress: to add limits for displayed cards when changing the width of  the screen.